// src/lib/generated/notification/openapi.gen.d.ts
export interface components {
  schemas: {
// ...
    PushPayload: {
      action_url?: string;
      badge?: string;
      body: string;
      campaign?: string;
      data?: {
        [key: string]: unknown;
      };
      icon?: string;
      image?: string;
      title: string;
      type: string;
    };
    SubscribeOutputBody: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/SubscribeOutputBody.json
       */
      readonly $schema?: string;
      data?: components["schemas"]["SubscribeResponse"];
      message: string;
      success: boolean;
    };
// ...
    TriggerByRoleIDsOutputBody: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/TriggerByRoleIDsOutputBody.json
       */
      readonly $schema?: string;
      message: string;
      success: boolean;
    };
    TriggerByRoleIDsRequest: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/TriggerByRoleIDsRequest.json
       */
      readonly $schema?: string;
      payload: components["schemas"]["PushPayload"];
      role_ids: number[] | null;
    };
    TriggerByUserIDsOutputBody: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/TriggerByUserIDsOutputBody.json
       */
      readonly $schema?: string;
      message: string;
      success: boolean;
    };
    TriggerByUserIDsRequest: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/TriggerByUserIDsRequest.json
       */
      readonly $schema?: string;
      payload: components["schemas"]["PushPayload"];
      user_ids: string[] | null;
    };
    UnsubscribeOutputBody: {
      /**
       * Format: uri
       * @description A URL to the JSON Schema for this object.
       * @example https://apisigma.iwkapps.com/notification/schemas/UnsubscribeOutputBody.json
       */
      readonly $schema?: string;
      message: string;
      success: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}